+++ src/lib.rs
@@ -383,8 +383,9 @@ pub use bdd::{
 };
 pub use cover::pla::{PLAWriter, PlaCover, PlaLabel};
 pub use cover::{
-    Anonymous, Cover, CoverType, Cube, CubeType, Disagreement, ExpandedMinterms, Label,
-    Minimizable, Minterm, MintermIter, NamedLabel, OutputSet, ReconcilableLabel, StringLabel,
+    Anonymous, Cover, CoverType, Cube, CubeType, Disagreement, ExpandedMinterms, FieldsIter,
+    InputField, Label, Minimizable, Minterm, MintermIter, NamedLabel, OutputSet, ReconcilableLabel,
+    StringLabel,
 };
 pub use espresso::{EspressoConfig, EspressoCubes};
 pub use expression::{BoolExpr, Expr, ExprBuilder, ExprNode, ExprVariables};